+++ ui/core/src/usecases/clp/index.ts
@@ -1,14 +1,9 @@
-import {
-  ErrorCode,
-  getErrorMessage,
-  IAsset,
-  IAssetAmount,
-} from "../../entities";
+import { IAsset } from "../../entities";
 import { UsecaseContext } from "..";
-import { PoolStore } from "../../store/pools";
 import { effect } from "@vue/reactivity";
-import { ReportTransactionError } from "../utils";
 import { Swap } from "./swap";
+import { AddLiquidity } from "./addLIquidity";
+import { RemoveLiquidity } from "./removeLiquidity";
 
 export default ({
   services,
@@ -17,10 +12,6 @@ export default ({
   "sif" | "clp" | "bus",
   "pools" | "wallet" | "accountpools"
 >) => {
-  const reportTransactionError = ReportTransactionError(services.bus);
-
-  const state = services.sif.getState();
-
   async function syncPools() {
     const state = services.sif.getState();
 
@@ -85,12 +76,6 @@ export default ({
     await syncPools();
   });
 
-  function findPool(pools: PoolStore, a: string, b: string) {
-    const key = [a, b].sort().join("_");
-
-    return pools[key] ?? null;
-  }
-
   effect(() => {
     // When sif address changes syncPools
     store.wallet.sif.address;
@@ -99,74 +84,8 @@ export default ({
 
   const actions = {
     swap: Swap(services),
-    async addLiquidity(
-      nativeAssetAmount: IAssetAmount,
-      externalAssetAmount: IAssetAmount,
-    ) {
-      if (!state.address) throw "No from address provided for swap";
-      const hasPool = !!findPool(
-        store.pools,
-        nativeAssetAmount.asset.symbol,
-        externalAssetAmount.asset.symbol,
-      );
-
-      const provideLiquidity = hasPool
-        ? services.clp.addLiquidity
-        : services.clp.createPool;
-
-      const tx = await provideLiquidity({
-        fromAddress: state.address,
-        nativeAssetAmount,
-        externalAssetAmount,
-      });
-
-      const txStatus = await services.sif.signAndBroadcast(tx.value.msg);
-
-      if (txStatus.state !== "accepted") {
-        // Edge case where we have run out of native balance and need to represent that
-        if (txStatus.code === ErrorCode.TX_FAILED_USER_NOT_ENOUGH_BALANCE) {
-          return reportTransactionError({
-            ...txStatus,
-            code: ErrorCode.TX_FAILED_NOT_ENOUGH_ROWAN_TO_COVER_GAS,
-            memo: getErrorMessage(
-              ErrorCode.TX_FAILED_NOT_ENOUGH_ROWAN_TO_COVER_GAS,
-            ),
-          });
-        }
-      }
-      return txStatus;
-    },
-
-    async removeLiquidity(
-      asset: IAsset,
-      wBasisPoints: string,
-      asymmetry: string,
-    ) {
-      const tx = await services.clp.removeLiquidity({
-        fromAddress: state.address,
-        asset,
-        asymmetry,
-        wBasisPoints,
-      });
-
-      const txStatus = await services.sif.signAndBroadcast(tx.value.msg);
-
-      if (txStatus.state !== "accepted") {
-        services.bus.dispatch({
-          type: "TransactionErrorEvent",
-          payload: {
-            txStatus,
-            message: txStatus.memo || "There was an error removing liquidity",
-          },
-        });
-      }
-
-      return txStatus;
-    },
-
-    async disconnect() {
-      services.sif.purgeClient();
-    },
+    addLiquidity: AddLiquidity(services, store),
+    removeLiquidity: RemoveLiquidity(services),
   };
 
   return actions;
